+++ packages/mrgn-ui/src/components/action-box-v2/actions/deposit-swap-box/components/action-input/components/lending-action/index.ts
@@ -1,1 +1,1 @@
-export * from "./swap-lend-action";
+export * from "./deposit-lend-action";
