+++ packages/mrgn-ui/src/components/action-box-v2/actions/deposit-swap-box/deposit-swap-box.tsx
@@ -239,7 +239,7 @@ export const DepositSwapBox = ({
   const buttonLabel = React.useMemo(() => (showCloseBalance ? "Close" : lendMode), [showCloseBalance, lendMode]);
 
   ///////////////////////
-  // Swap-Lend Actions //
+  // Deposit-Swap Actions //
   ///////////////////////
   const executeAction = async (
     params: ExecuteDepositSwapActionProps,
